move AnalysisSettings
