+++ biahub/estimate_stitch.py
@@ -6,7 +6,7 @@ import submitit
 
 from iohub import open_ome_zarr
 
-from biahub.analysis.AnalysisSettings import ProcessingSettings, StitchSettings
+from biahub.AnalysisSettings import ProcessingSettings, StitchSettings
 from biahub.cli.parsing import input_position_dirpaths, local, output_filepath
 from biahub.cli.utils import model_to_yaml
 from biahub.stitch import (
