+++ biahub/optimize_registration.py
@@ -5,7 +5,7 @@ import numpy as np
 
 from iohub import open_ome_zarr
 
-from biahub.analysis.AnalysisSettings import RegistrationSettings
+from biahub.AnalysisSettings import RegistrationSettings
 from biahub.cli.parsing import (
     config_filepath,
     output_filepath,
