+++ biahub/segment.py
@@ -8,7 +8,7 @@ import torch
 from iohub.ngff import open_ome_zarr
 from iohub.ngff.utils import create_empty_plate, process_single_position
 
-from biahub.analysis.AnalysisSettings import SegmentationSettings
+from biahub.AnalysisSettings import SegmentationSettings
 from biahub.cli import utils
 from biahub.cli.monitor import monitor_jobs
 from biahub.cli.parsing import (
